+++ CHANGES.rst
@@ -18,6 +18,16 @@ Bug fixes:
 - *add item here*
 
 
+1.2.12 (unreleased)
+-------------------
+
+Bug fixes:
+
+- purge transform cache when a uid referenced image
+  gets updated. this fixes `issue CMFPLone#2465 <https://github.com/plone/Products.CMFPlone/issues/2465>`_
+  [petschki]
+
+
 1.2.11 (2018-04-08)
 -------------------